Refractor: `MapList` MVVM 패턴화
